remade the registration forms
Complete HTML and CSS for new registration menu, with a more streamline method of doing it.

The JS of it still needs to be worked on, allowing for the menus to change more fluently but will work on this tomorrow.

Still to do:
link to DB, link everything to appropriate sites.